test: simplify e2e config
