+++ apps/examples/nextjs/.env.example
@@ -1,10 +1,8 @@
-HOST=127.0.0.1
-PORT=3001
+
 
 # Better Blog demo config
 # Use 'memory' for in-memory data, 'sql' for Postgres via DATABASE_URL
 BETTER_BLOG_PROVIDER=memory
 
 # Only used for sql provider
-SQL_DIALECT=postgres
 DATABASE_URL=postgres://user:pw@localhost:5432/better_blog
