+++ apps/examples/nextjs/next.config.ts
@@ -2,6 +2,7 @@ import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {
   // transpilePackages: ['better-blog']
+  //ignore scripts
   images: {
     remotePatterns: [
       {
