+++ apps/examples/nextjs/package.json
@@ -7,7 +7,8 @@
         "build": "next build --turbopack",
         "start": "next start",
         "start:e2e": "pnpm build && NODE_ENV=test next start",
-        "lint": "eslint"
+        "lint": "eslint",
+        "db:init": " pnpm -w exec dotenv -e apps/examples/nextjs/.env.local -- tsx apps/examples/nextjs/scripts/seed/kysley/init.ts"
     },
     "dependencies": {
         "@radix-ui/react-slot": "^1.2.3",
