+++ apps/examples/react/.env.example
@@ -1,13 +1,13 @@
+# React SPA configuration
+# Provider can be 'memory' for client-side only or 'api' for API backend
+VITE_BLOG_PROVIDER=memory
+
+# Only needed when VITE_BLOG_PROVIDER=api
+VITE_API_BASE_PATH=/api/posts
+VITE_API_BASE_URL=http://localhost:3005
+
+# Node.js API server config (only used with server.mjs)
+PORT=3005
 HOST=127.0.0.1
-PORT=3004
-
-# Better Blog demo config
-# Use 'memory' for in-memory data, 'api' to call the app's API routes
-BETTER_BLOG_PROVIDER=memory
-
-# Only used when BETTER_BLOG_PROVIDER=api
 API_BASE_PATH=/api/posts
-
-# Vite build-time mirrors (optional; the e2e runner maps these automatically if missing)
-VITE_BLOG_PROVIDER=
-VITE_API_BASE_PATH=
+BETTER_BLOG_PROVIDER=memory
