+++ apps/examples/react/.gitignore
@@ -9,6 +9,8 @@ lerna-debug.log*
 
 node_modules
 dist
+dist-memory
+dist-api
 dist-ssr
 *.local
