--- e2e/tests/global.teardown.react-memory.ts
@@ -1,19 +1,0 @@
-import { readFile } from 'node:fs/promises';
-import { join } from 'node:path';
-import { test as teardown } from '@playwright/test';
-
-const PROJECT = 'react-memory';
-
-teardown('cleanup react memory server', async () => {
-  try {
-    const metaPath = join('.e2e', `${PROJECT}.json`);
-    const metaRaw = await readFile(metaPath, 'utf8');
-    const { serverPid } = JSON.parse(metaRaw) as { serverPid?: number };
-    if (serverPid) {
-      try { process.kill(-serverPid); } catch {}
-      try { process.kill(serverPid); } catch {}
-    }
-  } catch {}
-});
-
-
