+++ package.json
@@ -15,10 +15,12 @@
         "@playwright/test": "^1.48.2",
         "@types/dockerode": "^3.3.26",
         "@types/pg": "^8.12.0",
+        "dockerode": "^4.0.8",
+        "dotenv-cli": "^10.0.0",
         "kysely": "^0.28.5",
         "pg": "^8.12.0",
-        "dockerode": "^4.0.8",
         "testcontainers": "^10.12.0",
+        "tsx": "^4.20.3",
         "turbo": "^2.5.5"
     },
     "pnpm": {
@@ -26,8 +28,14 @@
             "fumadocs-core": "15.7.11",
             "better-call": "1.0.19"
         },
-        "onlyBuiltDependencies": ["@biomejs/biome", "esbuild"],
-        "ignoredBuiltDependencies": ["@prisma/client", "@prisma/engines", "prisma"]
+        "onlyBuiltDependencies": [
+            "@biomejs/biome",
+            "esbuild"
+        ],
+        "ignoredBuiltDependencies": [
+            "@prisma/client",
+            "@prisma/engines",
+            "prisma"
+        ]
     }
 }
-
