+++ packages/better-blog/src/providers/sql/types.ts
@@ -1,5 +1,6 @@
 import type { BlogDataProviderConfig } from "@/types"
 import type { SQLDatabaseOptions } from "./dialect"
+export { DBSchema } from "./kysely-adapter"
 
 export type SQLProviderOptions = BlogDataProviderConfig & {
     database: SQLDatabaseOptions
